Crate resources_package

This crate allows you to package several files in your executable.

This is similar to include_bin! but easier to use when you have a lot of files.

Usage:

#![feature(phase)]
 
#[phase(plugin)]
extern crate resources_package;
extern crate resources_package_package;
 
static package: resources_package_package::Package = resources_package!([
    "path/to/resources",
    "other/path/to/other/resources"
]);

The type of the static variable is a resources_package_package::Package. See the documentation of resources_package_package.

Arguments

List of directories whose content is to be included.